addsub-ffmpeg/pyproject.toml
[project]
name = "addsub-ffmpeg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ffmpeg-python>=0.2.0",
]
